drv/user-leds: optionally blink some LEDs at start
We've been including the pong task in most basic/starter images for the
sole purpose of blinking some LEDs.

For the simplest case of a small board with a single debug LED and not a
lot of flash, this is quite wasteful (burns about a kiB). To conserve
resources on such boards, I've added a mechanism for configuring the
user-leds driver to start some LEDs blinking automatically.

Unlike with pong, the LEDs will all blink together, not in a sequence.
Such is the cost of frugality. On boards with one LED the distinction is
not material.

This imposes no size cost on existing uses of user-leds because the
configuration is evaluated at compile time, and will compile away. On
the oxcon2023 board it costs about 16 bytes to set the LED a-blinkin'.