hl::recv: handle message truncation
The hl::recv facility is _almost_ dead, but the single user I can find
(the I2C server) is quite popular, so we have to maintain it. Sigh.

It's entirely possible for rm.message_len to be larger than the buffer:
if the client sends an overly long message, the kernel will report that.
So, this code has been a client-controlled server panic for ages.

This switches it to check the message length when slicing, and transfer
the panic to the client, where it belongs.

This happens to remove a hundred bytes or so of text, by removing the
setup and call to the slice index bounds fail routine.