STM32G030/0xCon Demo Board I2C support.
This is the shotgun change required to teach all the various bits that
g030 exists, plus adding it to the oxcon2023g0 demo app.

The main functional change here: panic messages can now be disabled in
the i2c server for specific targets. It'd be nice to have explicit
control over this, instead of having it depend on the target, but the
way things are currently structured that's not practical.

With panic messages off we can fit the server in less than 4 kiB, making
it practical on STM32G030.