gimlet-seq-server: start driving fault pin.
This pulls the fault pin low (active) shortly after the sequencer task
first runs, and then sets it high (inactive) only once we get up to the
IPC loop.

Failures in the sequencer startup tend to cause panics, which will have
the effect of holding the pin low as the task is repeatedly restarted.
This also means we will be able to externally observe any other restart
of the task.