make Gimletlet use TOML inheritance (#1568)
This change reworks the various Gimletlet applications to utilize TOML inheritance. There is now a base-gimletlet2.toml which contains the core Hubris tasks required for the SP and declares the I2C and SPI buses. app-mgmt.toml utilizes a board type of gimletlet1 which lead me to make the distinction of gimletlet2 in the base TOML title.

The goal here was to reduce copy-paste duplication and also decouple the core gimletlet functionality from the various applications folks may have. Having these bits separated will make it easier for others to make custom gimletlet applications without worrying about legacy things that were committed to the old app.toml that was largely shared.