Stopped abusing SlotId in favor of treating the boot loader as a separate component.
Add Component* versions of APIs to allow selection of boot loader component.
Use Fwid type to represent the SHA3_256 digest in case we change digest algo someday.
Also, Fwid includes padding bytes and any subsequent programmed pages in a flash slot,
so don't just call it a digest which can be misleading to people who might expect that
a hash of the bytes of an image would match the RoT report.
Remove ImageVersion new structs to allow it to be deprecated and for the ImageHeader to be removed later.
Add versioned_rot_boot_info to select a particular RotBootInfo verison since the
existing rot_state and rot_boot_info need to remain stable for roll forward and roll back to
work when RoT version is newer than SP version.
Use conversion functions to isolate SP to MGS data structure conversion code.