hiffy:: Add `SendLeaseReadWrite` idol function behind 'send-rw' feature.
This function is called by `humility hiffy` to call hif operations that
require a read and a write lease. This commit hides this function behind
a new feature `send-rw` because including it in the existing `send`
feature causes the `hiffy` task in g0 images to roll over into another
flash block that isn't available on the donglet images. This new feature
is enabled by the existing `lpc55` and `stm32h7` features making this
mosly transparent to the app.tomls.