sprot: plumbing for attest & attest_len
This is the first op supported by sprot that takes both a read and a
write lease. Following the existing structure of `handle_request` we add
a new `Attest` variant to the TrailingData enum where we store the
appropriately sized slice that holds the read lease. As a result we must
be more explicit about the lifetimes in the `handle_request` method
because it's return value no longer has the same lifetime as the `self`
param (3rd lifetime elision rule).

The rest is pretty straight forward with the call to `Attest::attest`
happending in the match arm for this new `TrailingData` variant.