Update bitflags to 2.x
This brings us up to the most recent release, and has some nice
properties, like there now being a trait that describes the bitflags API
instead of it all being open-coded. (This is the property that motivated
this change.)

There are some slightly annoying aspects, like the operations in that
trait no longer being const (because we don't have const impls yet). So,
this required some surgery.