gimlet-seq: Die more politely (#1611)
Sleeping for 100ms in a loop is not a particularly efficient way for a
task to die permanently. Thanks to suggestions from @labbott and
@cbiffle, I've changed this so that the sequencer task now dies by
waiting for a notification with an empty notification mask, essentially
waiting forever for a notification that should never happen. This is
much more respectful of other tasks that might still be trying to do
stuff, and would like to be scheduled to do that stuff.

Also, I've cleaned up a couple other things based on some of Cliff's
suggestions. Because we no longer set the state when dying, we can pass
the `jefe` handle into `init` by value, which is a bit nicer.