ringbuf: get rid of separate `count_entry!` macro (#1621)
I didn't love that there were separate `count_entry!` and
`ringbuf_entry!` macros --- it would be nicer if adopting event counts
didn't require changing every macro invocation. This change reworks the
macros to use a single `ringbuf_entry!` macro, and dispatch whether or
not to count events based on whether the ringbuf was *declared* using
`counted_ringbuf!` or `ringbuf!`.

The way this works is by generating a module with the same name as the
ringbuf, and generating a function inside that module that actually
records the entry (and may or may not record event counts, depending on
the macro that declared the ringbuf). This way, `ringbuf_entry!` can
simply call `<RINGBUF_NAME>::record(...)`, and the function will
determine whether or not to count events.

It occurs to me that we could also use the module to generate the event
counter static without having to do a proc-macro-y concatenation of
idents.