ringbuf: remove `concat_idents!` proc-macros (#1621)
Since the previous commit changes the `ringbuf!` and `counted_ringbuf!`
macros to generate a module for each ringbuf, we can stick the counters
static in there. This obviates the need to use proc-macros for
incrementing the counters just so we can concatenate the ringbuf's name
with `_COUNTS` to determine the name of the counter static. Now, we only
use the proc macro for deriving `Count`. This allows us to make the
derive macro optional.