fix example breaking with `cargo test --workspace`
Currently, the `cargo test --workspace` invocation on CI tries to build
example code for the host target, which is nice, as ensuring that the
macro example builds at all is a useful test of the macros. However,
because `ringbuf` depends on crates that currently only build for ARM
targets, building the example for x86 breaks CI. To fix this, I've
changed the `build_util::expose_m_profile()` function to return a
`Result`, which is ignored when building lib crates like
`armv6m-atomic-hack` and `static-cell` on non-ARM targets, and
`unwrap`ped when building actual Hubris tasks. This way, it's now
possible to build the lib crates for the host target, so the example
compiles, but trying to compile Hubris for x86 will still fail
(correctly).