c-p-a: add counters to control-plane-agent (#1630)
Issue #1616 describes the need to add event counters to the
`control-plane-agent` task. In particular, it's desirable to have
counters of each IPC event and MGS message received by the
`control-plane-agent` task. This commit uses the ringbuf counters added
in #1621 and the `#[count(children)]` attribute added in this branch to
generate counters from `control-plane-agent`'s MGS message events.