gimlet_seq: count all IPC requests (#1630)
As described in #1616, it would be nice to have event counters for all
IPC messages received by the `gimlet-seq` and `control-plane-agent`
tasks. This branch adds a new `IpcRequest` variant to the `Trace` enum
in `gimlet-seq`, and uses it to generate counters of all IPC requests.
IPCs are now also included in the trace ringbuffer.