factor out counters into their own crate (#1630)
This commit moves the `Count` trait and `#[derive(Count)]` macro out of
`ringbuf` into a new `counters` crate, and adds a mechanism for
declaring a static set of counters _without_ an associated ring buffer.
This can be used in cases where we _just_ want to count stuff, like
error enums that don't have any additional data, or a total count of IPC
requests. I think @cbiffle wanted this for some stuff, so he'll be happy
to see this.

I've also changed the existing use of a separate ringbuf that's
fundamentally just for counters in `gimlet_seq` into an explicit
just-counters thingy, making it a bit smaller. I'll update Humility to
add a `counters` subcommand for reading counters regardless of whether
they're associated with a ringbuf, as well.