gimlet-seq: record sequencer T6 state on power off
As described in #1648, when the NIC fails to come up successfully and is
powered off, it's desirable to capture sequencer register state
associated with the NIC and stick them in the ringbuf before actually
commanding the sequencer to power off the NIC, in order to debug NIC
failures.

This branch changes the `gimlet-seq-server` task to record the values of
the `NIC_CTRL`, `NIC_STATUS`, `OUT_STATUS_NIC1` and `OUT_STATUS_NIC2`
registers in the ringbuf when `NIC_PWREN_L` is asserted high while in
the `A0PlusHP` state. Issue #1648 also requested that the `IFR` and
`AMD_STATUS` registers be recorded in the ringbuf when this occurs, but
those registers should already be present in the `Trace::Status` entry
that's recorded every time the sequencer task handles a notification,
immediately prior to checking if the task must transition the NIC's
state. So, I chose not to add additional ringbuf events for those
registers to avoid pushing other stuff off the end of the ringbuf.

Fixes #1648