gimlet-seq: record sequencer T6 state on power off (#1649)
As described in #1648, when the NIC fails to come up successfully and is
powered off, it's desirable to capture sequencer register state
associated with the NIC and stick them in the ringbuf before actually
commanding the sequencer to power off the NIC, in order to debug NIC
failures.

This branch changes the `gimlet-seq-server` task to record the values of
the `NIC_CTRL`, `NIC_STATUS`, `OUT_STATUS_NIC1` and `OUT_STATUS_NIC2`
registers in the ringbuf when `NIC_PWREN_L` is asserted high while in
the `A0PlusHP` state. Issue #1648 also requested that the `IFR` and
`AMD_STATUS` registers be recorded in the ringbuf when this occurs, but
those registers should already be present in the `Trace::Status` entry
that's recorded every time the sequencer task handles a notification,
immediately prior to checking if the task must transition the NIC's
state. So, I chose not to add additional ringbuf events for those
registers to avoid pushing other stuff off the end of the ringbuf.

As an example, here's the last few entries in the ringbuffer on 
Gimlet c12 when `t6init` comes up successfully:

```
  ... snipping out some stuff we don't care about ...

  38  815        1        1 A0(0x360)
  39  639        1        1 UpdateState(A0)
  40  507        1     9661 Status { ier: 0x0, ifr: 0x0, amd_status: 0x4, amd_a0: 0x4 }
  41  537        1        1 NICPowerEnableLow(false)
  42  639        1        1 UpdateState(A0PlusHP)
  43  507        1     5864 Status { ier: 0x0, ifr: 0x0, amd_status: 0x4, amd_a0: 0x4 }
  44  545        1        1 NICPowerEnableLow(true)
  45  551        1        1 NICStatus { nic_ctrl: 0x0, nic_status: 0x7f, out_status_nic1: 0x7f, out_status_nic2: 0x2 }
  46  639        1        1 UpdateState(A0)
  47  507        1       98 Status { ier: 0x0, ifr: 0x0, amd_status: 0x4, amd_a0: 0x4 }
  48  537        1        1 NICPowerEnableLow(false)
  49  639        1        1 UpdateState(A0PlusHP)
  50  507        1     1451 Status { ier: 0x0, ifr: 0x0, amd_status: 0x4, amd_a0: 0x4 }
```

Note that the `IFR` and `AMD_STATUS` regs are already included in the `Status`
entries.

Fixes #1648