stm32h7-nucleo: embiggen `sys` tasks
After #7632, the `sys` tasks in the STM32H7 Nucleo demo need a bit more
RAM to store counters. These apps no longer build with the current RAM
allocation, but #7635 was auto-merged because CI doesn't require the
builds for these devboard demo apps to pass in order to merge a PR.

This commit embiggens the RAM allocations for the `sys` tasks on these
boards. Now, they build again. My bad!