host-sp-comms: add event counters (#1638)
This branch adds event counters to `host-sp-comms`'s existing 20-element
ringbuffer, so that the total number of events can be tracked. In
particular, we count the total number of each message variant messages
sent and received between the host and SP, and each variant of the
`KeyLookupResult`, `KeySetResult`, and `InventoryResult` messages.

Depends on #1635.