spi: A reSPIte from excessive error codes (#1656)
Currently, the SPI API has a single big error enum which it returns from
all IPCs. This isn't all that great: for example, a `NothingToRelease`
error can be returned from the `Spi::read` IPC, which...doesn't make a
ton of sense. Additionally, many of these error variants are returned
only in the event of a programmer error in a client task, rather than
expected runtime errors, and probably should be `reply-fault`s rather
than error codes.

Per @cbiffle on Matrix:

> Re: the SPI errors, we've got
>
> - `TaskRestarted` - sure makes sense, also probably not appropriate to
>   unwrap though
> - `NothingToRelease` - should be reply-fault, indicates that you're
>   acting like you've locked the device when you haven't.
> - `BadDevice` - should be reply-fault, indicates that you've made up a
>   device index, that's what config is for
> - `BadTransferSize` - this one's subtle. The max transfer is
>   platform-specific, and there's currently no way to discover it. We
>   could treat this as a precondition and reply-fault it, but code that
>   wants to be diligent and check preconditions currently has no way to
>   do so! We could also define a common max value in the API and insist
>   upon it on all platforms, at which point this would be reply-fault.

Therefore, I've refactored the `SpiError` type to remove the
`NothingToRelease` and `BadDevice` variants, and changed the `Spi` IPC
interface so that the `lock` and `release` APIs only return runtime
errors on server death --- the `SpiError` type is now only returned by
`Spi::read`, `Spi::write`, and `Spi::exchange`. In situations where
`BadDevice` or `NothingToRelease` errors would previously have been
returned, the SPI server now sends a reply-fault instead.

The SPI API is a bit different from other driver APIs, as there's a
provision for running the SPI driver locally when a task has exclusive
ownership over a specific SPI bus, as well as running it as its own
driver task. In this case, the cases which would return reply-fault when
the driver is its own task now panic when the SPI interface's driver is
local to another task.

This reduces the total archive size for `gimlet/rev-f.toml` by 1%
(according to `cargo xtask sizes`), which is, honestly, not that much.
But, it still seems worthwhile.