ci: add test apps to `build-boards.yml` (#1667)
Currently, Hubris' CI doesn't build the tests at all. This means that
it's quite easy for the various test apps to bitrot --- for example, the
PSC test app doesn't build, because [it maps IRQs by bit, rather than by
name][1]. Named IRQ support was added back in 2022 (e.g. #424), so this
TOML clearly hasn't been touched in...kind of a while, which I think is
a pretty convincing argument for building these tests on CI!

[1]: https://github.com/oxidecomputer/hubris/blob/3325c6343a3e51a4d2c63cb3eac46085ba61dac3/test/tests-psc/app.toml#L68-L74