I2C: trace IRQ status on panic (#1666)
Using the `irq_status` system call I added in #1661, we can extend the
panic function added to `stm32xx-i2c` in #1657 to also include the state
of any IRQs currently mapped to its notification mask. This way, if an
I2C driver panics, we can also include the kernel's understanding of its
IRQ state in the dump.