kern: permit USlices to span contiguous MPU regions
Now that the build system has started packing tasks more cleverly, MPU
region boundaries can appear in the middle of task RAM and Flash in
positions that are difficult for the program to predict. This means the
program can't safely lend any section of RAM or Flash without risking
drawing the ire of the kernel, which --- owing to the limitations of
earlier versions of the build system --- would not accept any access
that spanned a region.

This fixes that. To make fixing that relatively inexpensive, this also
sorts task region tables by base address, so we can do a linear scan
with early exit.

See #1672 for more details and analysis.