Add the ability to update the RoT bootloader.
The `stage0next` contents is not copied to `stage0` unless its current hash matches
the hash of a valid signed image found at boot time.

Stage0 cache size is 8192 to accommodate previous 6000+ byte bootloaders.
That can be reduced when rollback protection is implemented and old bootloaders
are expunged from any spares and systems in use.