monorail-server: Add ringbuf counters
Depends on #1637

Now that the `vsc-err` types have `Count` implementations (as of #1637),
I thought it might be worthwhile to add counters to the
`monorail-server` ring bufs. This is especially nice because the one in
`main` only has two entries, and it seems worthwhile to be able to get a
total count of errors that have been observed.