net: Use `counters` for event counters, count errors (#1637)
Currently, the `net` task has some hand-implemented event counters.
Although these are simple and efficient, they don't match the format
that `humility counters` will look for when finding counters in a dump,
so they can't be queried using that command. Instead, they have to be
inspected using a series of `humility readvar` commands. This branch,
therefore, replaces the hand-written counters with the `counters` crate,
allowing the `net` task's event counters to be understood by `humility
counters`.

Additionally, I've added counters to the ringbufs in some of the BSP
modules (currently, only Gimletlet actually has ringbufs). For these, we
skip counting the various status and control register ringbuf entries
because they don't really seem to represent a countable event, just the
state of a register, and skipping all of them lets us avoid several
words of counters. Instead, we just count the number of error event
variants.